More precised calculations with MyPoint.java

+++ src/com/cookieso/function/MyPoint.java
@@ -1,8 +1,8 @@
 package com.cookieso.function;
 
 public class MyPoint {
-    public int x, y;
-    public MyPoint(int x, int y) {
+    public double x, y;
+    public MyPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
